ADDED: SCReenshOT - command line screen capture utility, keybinding.
scrot is a simple command line screen capture utility, it uses imlib2 to grab and save images.

scrot has many useful features:

Support for multiple image formats: JPG, PNG, GIF, and others.
The screenshot's quality is configurable.
It is possible to capture a specific window or a rectangular area on the screen.
Because scrot is a command line utility, it can easily be scripted and put to novel uses. For instance, scrot can be used to monitor an X server in absence.

scrot is free software under the MIT-advertising license.
https://github.com/resurrecting-open-source-projects/scrot.git